move to tokio for async filesystem operations and mutexes
Using a std mutex leads to deadblocks as the executor is unaware it
might not be able to make progress as the mutex is blocked. This causes
the executor to hang on that (especially when using only one thread).

Tokio comes with async aware mutexes and file system operations. It is
what the rust ecosystem has centralized on so let use it for everything.